nothing

+++ models/net.py
@@ -71,9 +71,9 @@ class network():
     def proposal_target_layer(self, rpn_rois, _gt_boxes, num_classes):
         rois, labels, bbox_targets, bbox_inside_weights, bbox_outside_weights = tf.py_func( proposal_target_layer_py,
                                                                                             [ rpn_rois, _gt_boxes, num_classes],
-                                                                                            [tf.float32, tf.int32, tf.float32, tf.float32, tf.float32])
+                                                                                            [tf.float32, tf.float32, tf.float32, tf.float32, tf.float32])
         rois = tf.reshape( rois, [-1, 5], name = 'rois')
-        labels = tf.convert_to_tensor( tf.cast(labels, tf.int32),name = 'labels')
+        labels = tf.convert_to_tensor( tf.cast(labels, tf.int32), name = 'labels')
         bbox_targets = tf.convert_to_tensor( bbox_targets, name = 'bbox_targets')
         bbox_inside_weights = tf.convert_to_tensor( bbox_inside_weights, name = 'bbox_inside_weights')
         bbox_outside_weights = tf.convert_to_tensor( bbox_outside_weights, name = 'bbox_outside_weights')
